cassandra-spark/requirements.txt
pymysql
cassandra-driver
mysql-connector-python